refactor: DashBoardScreenRootを作成し、PortStatusDetailActivityへの遷移を移動 不要な中間 DashboardScreen.kt は消した
